+++ packages/components/src/Aside/Aside.test.tsx
@@ -1,6 +1,5 @@
 /* eslint-disable testing-library/no-node-access */
 import { render, screen } from '@testing-library/react';
-import React from 'react';
 import { Aside } from './Aside';
 
 test('default spacing is "none"', () => {