+++ packages/components/src/Breadcrumbs/BreadcrumbEllipsis.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import { ComponentProps, ReactNode, isValidElement } from 'react';
 import { Menu, MenuItem, MenuTrigger } from 'react-aria-components';
 import { useLocalizedStringFormatter } from '@react-aria/i18n';
 import { useClassNames } from '@marigold/system';
@@ -7,8 +7,8 @@ import { Popover } from '../Overlay/Popover';
 import { intlMessages } from '../intl/messages';
 import { BreadcrumbsItemProps } from './BreadcrumbsItem';
 
-interface BreadcrumbEllipsisProps extends React.ComponentProps<'span'> {
-  hiddenItems?: React.ReactNode[];
+interface BreadcrumbEllipsisProps extends ComponentProps<'span'> {
+  hiddenItems?: ReactNode[];
   disabled?: boolean;
 }
 
@@ -29,7 +29,7 @@ export const BreadcrumbEllipsis = ({
       <Popover>
         <Menu className={container}>
           {hiddenItems.map((item, index) => {
-            if (!React.isValidElement<BreadcrumbsItemProps>(item)) return null;
+            if (!isValidElement<BreadcrumbsItemProps>(item)) return null;
 
             const { href, children: itemChildren } = item.props;