+++ packages/components/src/FieldBase/FieldBase.test.tsx
@@ -1,6 +1,5 @@
 import { screen } from '@testing-library/react';
 import type { ReactNode } from 'react';
-import React from 'react';
 import { TextField } from 'react-aria-components';
 import { Theme, cva } from '@marigold/system';
 import { setup } from '../test.utils';